Add gmp dep for Linux
It appears we end up linking with gmp. I suppose this is because Haskell
uses gmp on Linux and thus we link to a system installed gmp. libgmp-dev
is pulled in with the install of Haskell but since actonc might run on a
system where Haskell is not installed, there is no guarantee libgmp will
be present. We must state explicitly that we depend on gmp!

Closes #7.

Signed-off-by: Kristian Larsson <841628+plajjan@users.noreply.github.com>