Add workflow to auto-merge updates
Updates to formula (just the one acton formula in this case) are created
through a new branch and a PR. When the test workflow has passed on that
branch, the pr-pull label can be added to the PR which will trigger the
next workflow that merges in the work to the main branch. It's not a
classic merge as the formula needs to be updated with sha sums etc from
the binary artifacts produced by the test build.

Once we see a PR complete its test workflow successfully, we want to
automatically add the pr-pull label so the work can continue without
human intervention. We achieve that with yet another workflow.