Turn off matrix fail-fast
This will allow us to see the build output of the other builds even if
one fails in the matrix. I think this will be useful. Quite a lot of the
time the linux build fails first, because it runs faster, but then we
have no idea what the build result is like on macos because those builds
were canceled.

Closes #21.

Signed-off-by: Kristian Larsson <841628+plajjan@users.noreply.github.com>