Fix triggering of workflow on label
The pr-pull workflow is triggered by adding the pr-pull label. However,
in order to prevent endless loops by recursively calling a workflow from
another workflow, Github Actions are designed such that using the
standard github.token will never lead to triggering another workflow.
This is a design feature, removing a potential footgun.

However, since we actually want to trigger the pr-pull workflow we have
to use a different token for authentication when adding the tag!